workspace/python/agents/math-agent-otel/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "math-agent-otel"
version = "0.1.0"
description = "具有 OpenTelemetry 儀表化和 Jaeger 追蹤功能的 ADK 數學代理"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "ADK Team", email = "adk-team@google.com" }]
dependencies = [
    "google-adk>=1.15.0",
    "google-genai>=1.15.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=0.45b0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.23.0", "pytest-cov>=4.1.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["math_agent*"]
